% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_you_mean.R
\name{did_you_mean}
\alias{did_you_mean}
\title{String Mismatch Handling}
\usage{
did_you_mean(x, y, max_matches = 5)
}
\arguments{
\item{x}{A character vector. The object to search in.}

\item{y}{A character vector. The object to search for.}

\item{max_matches}{An integer. Maximum number of suggestions to print.}
}
\value{
invisible NULL
}
\description{
Generate errors for mismatches between strings. Closest matches by
Levenshtein distance are calculated to offer suggestions in error message.
}
\keyword{internal}
