% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist}
\alias{dist}
\title{Calculate Distance/Dissimilarity Metric}
\usage{
dist(
  obj,
  dis = "bray",
  part.weights = "uniform",
  ilr.weights = "uniform",
  pseudocount = 1e-12,
  normalized = TRUE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{obj}{A phyloseq object.}

\item{dis}{Name of distance to use. One of philr, u.unifrac, w.unifrac,
or a metric accepted by the method argument in \link[vegan]{vegdist}.}

\item{part.weights}{See \link[philr]{philr}.}

\item{ilr.weights}{See \link[philr]{philr}.}

\item{pseudocount}{Passed to \link[vegan]{vegdist} if dis is aitchison or
robust.aitchison, or added to ALL counts of the feature table if
dis = philr.}

\item{normalized}{See \link[phyloseq]{UniFrac}.}

\item{parallel}{See \link[phyloseq]{UniFrac}.}

\item{...}{Passed to \link[vegan]{vegdist} and \link[philr]{philr}.}
}
\value{
A distance matrix.
}
\description{
Compute a distance matrix from the feature table of a phyloseq object.
}
\details{
\strong{SAMPLES ARE ROWS}
}
\keyword{internal}
