% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoaSH.R
\name{pcoaSH}
\alias{pcoaSH}
\title{PCoA Overview}
\usage{
pcoaSH(obj)
}
\arguments{
\item{obj}{A _phyloseq_ object.}
}
\value{
Starts a Shiny App.
}
\description{
View PCoA plots in an interactive shiny app. Select distance/dissimilarity
metrices and subset data.
}
\examples{
pcoaSH(GlobalPatterns)
}
\keyword{MBT}
