% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{Plot QC Summaries}
\usage{
plotQC(
  data,
  option,
  meta_data_col,
  fill.log = FALSE,
  add.mean = TRUE,
  q = 3,
  qs = c(0.09, 0.25, 0.5, 0.75, 0.91)
)
}
\arguments{
\item{data}{A list returned by \code{readQC()}}

\item{option}{One of c("1","2","3","4","5").
\itemize{
\item "1": Quantile plot
\item "2": Mean quality per cycle.
\item "3": Read count per sample.
\item "4": Quantile boxplot (qiime2.view)
\item "5": Quality heatmap
}}

\item{meta_data_col}{Character specifying the the metadata column to color read couts by.}

\item{fill.log}{A boolean. Scale fill of option "5"}

\item{add.mean}{A boolean. Plot means on top of heatmap ("5")}

\item{q}{Integer. One of 1:length(qs)}

\item{qs}{Numeric vector. Specifies quantiles to calculate.}
}
\value{
A ggplot object.
}
\description{
Plot results of \code{readQC}.
}
\examples{
plotQC(data)
}
\keyword{MBT}
