% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefySH.R
\name{rarefySH}
\alias{rarefySH}
\title{Rarefaction Overview}
\usage{
rarefySH(obj)
}
\arguments{
\item{obj}{A \emph{phyloseq} object.}
}
\value{
Starts a Shiny App.
}
\description{
Interactively assess effects of rarefaction depth on features and samples.
}
\examples{
rarefySH(GlobalPatterns)
}
\keyword{MBT}
