% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQC.R
\name{readQC}
\alias{readQC}
\title{Generate QC Summaries}
\usage{
readQC(
  file_paths,
  qs = c(0.09, 0.25, 0.5, 0.75, 0.91),
  meta_data,
  id_col,
  extract_pattern,
  rm_pattern = ""
)
}
\arguments{
\item{file_paths}{A character vector of relative or absolute paths of
fastq files.}

\item{qs}{A numeric vector of quantiles of PHRED scores to compute.}

\item{meta_data}{A dataframe storing metadata.}

\item{id_col}{Sample identifier column in metadata.}

\item{extract_pattern}{A string specifying the sample identifier in the
file name. See details.}

\item{rm_pattern}{A string specifying the remainder to remove. See details.}
}
\value{
A List. Elements are:
\itemize{
\item "read.counts": A dataframe of read counts per file and respective
metadata.
\item "read.means": A dataframe of mean PHRED scores per cycle.
\item "read.quants": A (molten) dataframe of PHRED scores per cycle and
quantile.
\item "quality.tiles": A (molten) dataframe of counts per PHRED score and cycle.
\item "max.read.length": Integer.
}
}
\description{
Summarize quality information of one or multiple fastq files.
}
\details{
Sample identifiers are assumed to be a substring of the respective
filenames. \strong{extract_pattern} is used to match and extract this pattern
from file names. Since it might be necessary to match field
delimiters as well, \strong{rm_pattern} can be specified to remove a pattern from
sample identifiers. Both patterns accept regular expressions.
}
\examples{
readQC(c("NGSXX_041_lib123_1_.fastq",
"NGSXX_042_lib123_1_.fastq"), meta.dat, "SampleID", "_\\\\d\\\\d\\\\d_", "_")
}
\keyword{MBT}
