% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_progress.R
\name{show_progress}
\alias{show_progress}
\title{Show Loop Progress}
\usage{
show_progress(i, n, names)
}
\arguments{
\item{i}{An integer. The loop counter.}

\item{n}{An integer. The maximum loop count.}

\item{names}{A character vector. The names of the dimension being looped
over, e.g. row.names.}
}
\description{
Print progress bar, percentage, and current iteration to console.
}
\keyword{internal}
