% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaSH.R
\name{taxaSH}
\alias{taxaSH}
\title{Taxa Overview}
\usage{
taxaSH(obj, plot_height = NA)
}
\arguments{
\item{obj}{A \emph{phyloseq} object.}

\item{plot_height}{An integer number specifying the plot height in \emph{px}.
Defaults to 1.6 times the size of a default plot window opened from Rscript.}
}
\description{
Start a \emph{Shiny App} to explore taxonomic composition of a \emph{phyloseq} dataset.
}
\examples{
taxaSH(GlobalPatterns)
}
\keyword{MBT}
