% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist}
\alias{dist}
\title{Distance Metric}
\usage{
dist(
  obj,
  dis = "bray",
  part.weights = "uniform",
  ilr.weights = "uniform",
  pseudocount = 1e-12,
  normalized = TRUE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{obj}{A phyloseq object.}

\item{dis}{Name of distance to use. One of _philr_, _u.unifrac_, _w.unifrac_,
or a metric accepted by the _method_ argument in *vegan::vegdist()*.}

\item{part.weights}{See *philr::philr()*.}

\item{ilr.weights}{See *philr::philr()*.}

\item{pseudocount}{Passed to *vegan::vegdist()* if dis is _aitchison_ or
_robust.aitchison_, or added to ALL counts of feauture table if
_dis = philr_.}

\item{normalized}{See *phyloseq::UniFrac()*.}

\item{parallel}{See *phyloseq::UniFrac()*.}

\item{...}{Passed to *vegan::vegdist()* and *philr::philr()*.}
}
\value{
distance matrix
}
\description{
Compute distances of the feature table of a _phyloseq_ object.
}
\examples{
mbt_dist <- MBT::dist(GlobalPatterns, dis = w.unifrac)
}
\keyword{MBT}
