% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{Plot QC Summaries}
\usage{
plotQC(
  data,
  option,
  meta_data_col,
  fill.log = FALSE,
  add.mean = TRUE,
  q = 3,
  qs = c(0.09, 0.25, 0.5, 0.75, 0.91),
  ...
)
}
\arguments{
\item{data}{A list returned by readQC().}

\item{option}{One of:
\itemize{
\item "qp": Quantile plot
\item "mq": Mean quality per cycle.
\item "rc": Read count per sample.
\item "bp": Quantile boxplot (similar to quality plot in qzv artifact produced
by \strong{qiime demux summarize})
\item "hm": Quality heatmap
}}

\item{meta_data_col}{Character specifying the the metadata column to color
read counts by.}

\item{fill.log}{A boolean. Scale fill of option "hm".}

\item{add.mean}{A boolean. Plot means on top of heatmap ("hm")}

\item{q}{Integer. One of 1:length(qs)}

\item{qs}{Numeric vector. Specifies quantiles calculated in readQC().}

\item{...}{Passed to \link[viridis]{scale_fill_viridis_c}}
}
\value{
A ggplot object.
}
\description{
Various plotting options for results of readQC.
}
\examples{
plotQC(data)
}
\keyword{MBT}
