% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQC.R
\name{readQC}
\alias{readQC}
\title{Generate QC Summaries}
\usage{
readQC(
  file_paths,
  qs = c(0.09, 0.25, 0.5, 0.75, 0.91),
  meta_data,
  id_col,
  extract_pattern,
  rm_pattern
)
}
\arguments{
\item{file_paths}{A character vector of relative or absolute paths of fastq files.}

\item{qs}{A numeric vector of quantiles of PHRED scores to compute.}

\item{meta_data}{A dataframe storing metadata.}

\item{id_col}{Metadata column name where sample identifiers are stored.}

\item{extract_pattern}{A string specifying the sample id in the file name.}

\item{rm_pattern}{A string specifying the remainder to remove.}
}
\value{
List of length 5.
}
\description{
Summarize quality aspects and features of FASTQ files.
}
\examples{
readQC("example.fastq", meta_data = meta.data, id_col = "sample_name", extract_pattern = "\\\\.\\\\d\\\\d\\\\d\\\\.", rm_pattern = "\\\\.")
}
\keyword{MBT}
