% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaSH.R
\name{taxaSH}
\alias{taxaSH}
\title{Taxa Overview}
\usage{
taxaSH(obj, plot_height = NA)
}
\arguments{
\item{obj}{A _phyloseq_ object.}

\item{plot_height}{An integer number specifying the plot height in *px*.
Defaults to 1.6 times the size of a default plot window opened from Rscript.}
}
\description{
Start a *Shiny App* to explore taxonomic composition of a _phyloseq_ dataset.
}
\examples{
taxaSH(GlobalPatterns)
}
\keyword{MBT}
